use existed DEM and trans files in working dir while loading
move is_file_exist() and check_loaded_dataset() from pysarApp.py to _pysar_utilities.py

load_data.py:
Check existed DEM and mapping transformation files in PYSAR folder, if input file path is None or auto

pysarApp.py:
add --reset option to reset files attribute for a fresh pysarApp re-run.

plot_pysarApp.sh
add ploting for DEM error file.
